STOR-1/Organise project files
Organise a repository to make it easier to develop, maintain, and collaborate
