+++ app/templates/accounts/confirm_email.html
